+++ proddaturiMinApp/src/main/java/com/org/proddaturiMinApp/controller/CategoryAdminController.java
@@ -1,11 +1,10 @@
 package com.org.proddaturiMinApp.controller;
 
 import com.org.proddaturiMinApp.model.Category;
-import com.org.proddaturiMinApp.service.CategoryService;
+import com.org.proddaturiMinApp.service.impl.CategoryServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Optional;
 
 
@@ -14,7 +13,7 @@ import java.util.Optional;
 
 public class CategoryAdminController {
     @Autowired
-    private CategoryService categoryService;
+    private CategoryServiceImpl categoryService;
 
     @PostMapping("/addCategory")
     public String addCategory(@RequestBody Category category) {
